+++ src/DateTimeExtensions/WorkingDays/CultureStrategies/DE_ATHolidayStrategy.cs
@@ -24,6 +24,34 @@ namespace DateTimeExtensions.WorkingDays.CultureStrategies
             this.InnerCalendarDays.Add(new Holiday(ChristianHolidays.Christmas));
             this.InnerCalendarDays.Add(new Holiday(ChristianHolidays.StStephansDay));
             this.InnerCalendarDays.Add(new Holiday(GlobalHolidays.NewYearsEve));
+            this.InnerHolidays.Add(GlobalHolidays.NewYear);
+            this.InnerHolidays.Add(ChristianHolidays.Epiphany);
+            this.InnerHolidays.Add(ChristianHolidays.EasterMonday);
+            this.InnerHolidays.Add(GlobalHolidays.InternationalWorkersDay);
+            this.InnerHolidays.Add(ChristianHolidays.Ascension);
+            this.InnerHolidays.Add(ChristianHolidays.PentecostMonday);
+            this.InnerHolidays.Add(ChristianHolidays.CorpusChristi);
+            this.InnerHolidays.Add(ChristianHolidays.Assumption);
+            this.InnerHolidays.Add(AustrianNationalHoliday);
+            this.InnerHolidays.Add(ChristianHolidays.AllSaints);
+            this.InnerHolidays.Add(ChristianHolidays.ImaculateConception);
+            this.InnerHolidays.Add(ChristianHolidays.Christmas);
+            this.InnerHolidays.Add(ChristianHolidays.StStephansDay);
+            this.InnerHolidays.Add(GlobalHolidays.NewYearsEve);
+        }
+
+        private static Holiday austrianNationalHoliday;
+
+        public static Holiday AustrianNationalHoliday
+        {
+            get
+            {
+                if (austrianNationalHoliday == null)
+                {
+                    austrianNationalHoliday = new FixedHoliday("Austrian National Holiday", 10, 26);
+                }
+                return austrianNationalHoliday;
+            }
         }
             
         public static NamedDayInitializer AustrianNationalHoliday { get; } = new NamedDayInitializer(() =>