+++ src/DateTimeExtensions/WorkingDays/CultureStrategies/IT_ITHolidayStrategy.cs
@@ -59,3 +59,70 @@
 //         }
 //     }
 // }
+=======
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// 
+
+#endregion
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using DateTimeExtensions.Common;
+
+namespace DateTimeExtensions.WorkingDays.CultureStrategies
+{
+    [Locale("it-IT")]
+    public class IT_ITHolidayStrategy : HolidayStrategyBase, IHolidayStrategy
+    {
+        public IT_ITHolidayStrategy()
+        {
+            this.InnerHolidays.Add(GlobalHolidays.NewYear);
+            this.InnerHolidays.Add(ChristianHolidays.Epiphany);
+            this.InnerHolidays.Add(ChristianHolidays.EasterMonday);
+            this.InnerHolidays.Add(LiberationDay);
+            this.InnerHolidays.Add(GlobalHolidays.InternationalWorkersDay);
+            this.InnerHolidays.Add(RepublicDay);
+            this.InnerHolidays.Add(ChristianHolidays.Assumption);
+            this.InnerHolidays.Add(ChristianHolidays.AllSaints);
+            this.InnerHolidays.Add(ChristianHolidays.ImaculateConception);
+            this.InnerHolidays.Add(ChristianHolidays.Christmas);
+            this.InnerHolidays.Add(ChristianHolidays.StStephansDay);
+        }
+
+        //2 June - Republic Day
+        private static Holiday republicDay;
+
+        public static Holiday RepublicDay
+        {
+            get
+            {
+                if (republicDay == null)
+                {
+                    republicDay = new FixedHoliday("Republic Day", 6, 2);
+                }
+                return republicDay;
+            }
+        }
+        
+        //25 April - Liberation Day
+        private static Holiday liberationDay;
+
+        public static Holiday LiberationDay
+        {
+            get
+            {
+                if (liberationDay == null)
+                {
+                    liberationDay = new FixedHoliday("Liberation Day", 4, 25);
+                }
+                return liberationDay;
+            }
+        }
+    }
+}